feat: apis 전역으로 통합
