+++ src/components/modal/archivingModal/editArchivingModal/EditArchivingModal.tsx
@@ -15,6 +15,7 @@ import Modal from 'react-native-modal'
 import { useMutation } from 'react-query'
 import { useRecoilValue } from 'recoil'
 
+import { patchArchiving } from '@/apis/archiving'
 import { defaultIcons, defaultImages } from '@/assets'
 import { BoxButton } from '@/components/buttons/boxButton/BoxButton'
 import { DropDown } from '@/components/dropDown/DropDown'
@@ -39,7 +40,6 @@ import {
   TextInput,
   Title,
 } from '../ArchivingModal.style'
-import { patchArchiving } from '../apis/archiving'
 
 interface EditArchivingModalProps {
   archivingId: number
