+++ src/screens/upload/linkUpload/LinkUpload.tsx
@@ -5,6 +5,7 @@ import { KeyboardAvoidingView, Platform, ScrollView, Text } from 'react-native'
 import { useMutation } from 'react-query'
 import { useRecoilState, useRecoilValue } from 'recoil'
 
+import { postContents } from '@/apis/content'
 import { BoxButton } from '@/components/buttons/boxButton/BoxButton'
 import { CloseButtonHeader } from '@/components/headers/closeButtonHeader/CloseButtonHeader'
 import { SelectArchivingModal } from '@/components/modal/selectArchivingModal/SelectArchivingModal'
@@ -25,7 +26,6 @@ import {
   TextInput,
   Title,
 } from '../Upload.style'
-import { postContents } from '../apis/postContents'
 
 /**
  *
